stream: re-use legacy destroyer
PR-URL: https://github.com/nodejs/node/pull/31316
