+++ lib/internal/streams/async_iterator.js
@@ -11,6 +11,7 @@ const {
 } = primordials;
 
 const finished = require('internal/streams/end-of-stream');
+const destroyImpl = require('internal/streams/destroy');
 
 const kLastResolve = Symbol('lastResolve');
 const kLastReject = Symbol('lastReject');
@@ -21,15 +22,6 @@ const kHandlePromise = Symbol('handlePromise');
 const kStream = Symbol('stream');
 
 let Readable;
-
-function destroy(stream, err) {
-  // request.destroy just do .end - .abort is what we want
-  if (typeof stream.abort === 'function') return stream.abort();
-  if (stream.req &&
-      typeof stream.req.abort === 'function') return stream.req.abort();
-  if (typeof stream.destroy === 'function') return stream.destroy(err);
-  if (typeof stream.close === 'function') return stream.close();
-}
 
 function createIterResult(value, done) {
   return { value, done };
@@ -92,7 +84,7 @@ function finish(self, err) {
         resolve(createIterResult(undefined, true));
       }
     });
-    destroy(stream, err);
+    destroyImpl.destroyer(stream, err);
   });
 }
 
@@ -174,7 +166,7 @@ const createReadableStreamAsyncIterator = (stream) => {
 
     const src = stream;
     stream = new Readable({ objectMode: true }).wrap(src);
-    finished(stream, (err) => destroy(src, err));
+    finished(stream, (err) => destroyImpl.destroyer(src, err));
   }
 
   const iterator = ObjectCreate(ReadableStreamAsyncIteratorPrototype, {
