+++ lib/internal/streams/pipeline.js
@@ -12,6 +12,7 @@ const {
 let eos;
 
 const { once } = require('internal/util');
+const destroyImpl = require('internal/streams/destroy');
 const {
   ERR_INVALID_ARG_TYPE,
   ERR_INVALID_RETURN_VALUE,
@@ -23,10 +24,6 @@ const {
 let EE;
 let PassThrough;
 let createReadableStreamAsyncIterator;
-
-function isRequest(stream) {
-  return stream && stream.setHeader && typeof stream.abort === 'function';
-}
 
 function destroyer(stream, reading, writing, callback) {
   callback = once(callback);
@@ -49,11 +46,7 @@ function destroyer(stream, reading, writing, callback) {
     if (destroyed) return;
     destroyed = true;
 
-    // request.destroy just do .end - .abort is what we want
-    if (isRequest(stream)) return stream.abort();
-    if (isRequest(stream.req)) return stream.req.abort();
-    if (typeof stream.destroy === 'function') return stream.destroy(err);
-    if (typeof stream.close === 'function') return stream.close();
+    destroyImpl.destroyer(stream, err);
 
     callback(err || new ERR_STREAM_DESTROYED('pipe'));
   };
